Always set the MAC address of wlan0
The primary use of jenny is in a virtual machine and not on a Raspberry
Pi with a wlan0 network device. Consequently we can't determine and
tweak the MAC address of the wlan0 network device as commented in
bin/jenny.

Instead, simpy set the MAC address of the wlan0 network device to the
pre-determined when it is brought up.