chore: interface -> internal interface
Until now, wlan0 has been referred to as "the interface" or "the ingress
interface". The former is confusing because jenny (and pinc, by
extension) relies on the system having two interfaces; the latter is
confusing because traffic ingresses and egresses both of those
interfaces.

To balance the EXTERNAL_INTERFACE constant introduced in e58eac68,
rename all instances of interface to internal interface where sensible.